moved event manager to outside package
