+++ node/peer.go
@@ -1,0 +1,34 @@
+// Copyright (c) 2017 The Alvalor Authors
+//
+// This file is part of Alvalor.
+//
+// Alvalor is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Alvalor is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with Alvalor.  If not, see <http://www.gnu.org/licenses/>.
+
+package node
+
+import (
+	"github.com/alvalor/alvalor-go/node/peer"
+	"github.com/alvalor/alvalor-go/types"
+)
+
+type Peers interface {
+	Active(address string)
+	Inactive(address string)
+	Requested(address string, hash types.Hash)
+	Received(address string, hash types.Hash)
+	Pending(address string) ([]types.Hash, error)
+	Seen(address string) ([]types.Hash, error)
+	Addresses(filters ...peer.FilterFunc) []string
+	Count(filters ...peer.FilterFunc) uint
+}
